Changed semaphore in comm_lend-light for atomic calls --> changed compiler from xl to gcc
Removed warnings when compiling

Added paraver events when tracing (without crashing when not tracig)

dlb/dlb_API.c --> Changed the way of measuring time not dependent from OpenMP

+++ LB_MPI/MPI_interface.c
@@ -2,6 +2,7 @@
 #include <process_MPI.h>
 #include <mpi.h>
 #include <MPI_calls_coded.h>
+#include <stdint.h>
 
 void DLB_MPI_Init_enter (int *argc, char ***argv){
 	#ifdef debugInOut
@@ -36,7 +37,7 @@ void DLB_MPI_Sendrecv_enter ( void *sendbuf, int sendcount, MPI_Datatype sendtyp
 	fprintf(stderr," >> MPI_Sendrecv...............\n");
 	#endif
 	
-	before_mpi(SendRecv, (int)sendbuf, dest);
+	before_mpi(SendRecv, (intptr_t)sendbuf, dest);
 }
 
 void DLB_MPI_Sendrecv_leave (){
@@ -52,7 +53,7 @@ void DLB_MPI_Sendrecv_replace_enter ( void *buf, int count, MPI_Datatype datatyp
 	fprintf(stderr," >> MPI_Sendrecv_replace...............\n");
 	#endif
 	
-	before_mpi(SendRecv, (int)buf, dest);
+	before_mpi(SendRecv, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Sendrecv_replace_leave (){
@@ -68,7 +69,7 @@ void DLB_MPI_Bsend_enter (void* buf, int count, MPI_Datatype datatype, int dest,
 	fprintf(stderr," >> MPI_Bsend...............\n");
 	#endif
 	
-	before_mpi(Bsend, (int)buf, dest);
+	before_mpi(Bsend, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Bsend_leave (){
@@ -84,7 +85,7 @@ void DLB_MPI_Ssend_enter (void*buf, int count, MPI_Datatype datatype, int dest, 
 	fprintf(stderr," >> MPI_Ssend...............\n");
 	#endif
 	
-	before_mpi(Ssend, (int)buf, dest);
+	before_mpi(Ssend, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Ssend_leave (){
@@ -100,7 +101,7 @@ void DLB_MPI_Rsend_enter (void*buf, int count, MPI_Datatype datatype, int dest, 
 	fprintf(stderr," >> MPI_Rsend...............\n");
 	#endif
 	
-	before_mpi(Rsend, (int)buf, dest);
+	before_mpi(Rsend, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Rsend_leave (){
@@ -116,7 +117,7 @@ void DLB_MPI_Send_enter (void*buf, int count, MPI_Datatype datatype, int dest, i
 	fprintf(stderr," >> MPI_Send...............\n");
 	#endif
 	
-	before_mpi(Send, (int)buf, dest);
+	before_mpi(Send, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Send_leave (){
@@ -132,7 +133,7 @@ void DLB_MPI_Ibsend_enter (void *buf, int count, MPI_Datatype datatype, int dest
 	fprintf(stderr," >> MPI_Ibsend...............\n");
 	#endif
 
-	before_mpi(Ibsend, (int)buf, dest);	
+	before_mpi(Ibsend, (intptr_t)buf, dest);	
 }
 
 void DLB_MPI_Ibsend_leave (){
@@ -148,7 +149,7 @@ void DLB_MPI_Isend_enter (void*buf, int count, MPI_Datatype datatype, int dest, 
 	fprintf(stderr," >> MPI_Isend...............\n");
 	#endif
 	
-	before_mpi(Isend, (int)buf, dest);
+	before_mpi(Isend, (intptr_t)buf, dest);
 
 }
 
@@ -165,7 +166,7 @@ void DLB_MPI_Issend_enter (void*buf, int count, MPI_Datatype datatype, int dest,
 	fprintf(stderr," >> MPI_Issend...............\n");
 	#endif
 
-	before_mpi(Issend, (int)buf, dest);
+	before_mpi(Issend, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Issend_leave (){
@@ -181,7 +182,7 @@ void DLB_MPI_Irsend_enter (void*buf, int count, MPI_Datatype datatype, int dest,
 	fprintf(stderr," >> MPI_Irsend...............\n");
 	#endif
 
-	before_mpi(Irsend, (int)buf, dest);
+	before_mpi(Irsend, (intptr_t)buf, dest);
 }
 
 void DLB_MPI_Irsend_leave (){
@@ -197,7 +198,7 @@ void DLB_MPI_Recv_enter (void*buf, int count, MPI_Datatype datatype, int source,
 	fprintf(stderr," >> MPI_Recv...............\n");
 	#endif
 	
-	before_mpi(Recv, (int)buf, source);
+	before_mpi(Recv, (intptr_t)buf, source);
 }
 
 void DLB_MPI_Recv_leave (){
@@ -213,7 +214,7 @@ void DLB_MPI_Irecv_enter (void* buf, int count, MPI_Datatype datatype, int sourc
 	fprintf(stderr," >> MPI_Irecv...............\n");
 	#endif
 
-	before_mpi(Irecv, (int)buf, source);
+	before_mpi(Irecv, (intptr_t)buf, source);
 }
 
 void DLB_MPI_Irecv_leave (){
@@ -229,7 +230,7 @@ void DLB_MPI_Reduce_enter (void* sendbuf, void* recvbuf, int count, MPI_Datatype
 	fprintf(stderr," >> MPI_Reduce...............\n");
 	#endif
 	
-	before_mpi(Reduce, (int)sendbuf, root);
+	before_mpi(Reduce, (intptr_t)sendbuf, root);
 }
 
 void DLB_MPI_Reduce_leave (){
@@ -245,7 +246,7 @@ void DLB_MPI_Reduce_scatter_enter (void* sendbuf, void* recvbuf, int* recvcounts
 	fprintf(stderr," >> MPI_Reduce_scatter...............\n");
 	#endif
 	
-	before_mpi(Reduce_scatter, (int)sendbuf, (int)recvbuf);
+	before_mpi(Reduce_scatter, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Reduce_scatter_leave (){
@@ -261,7 +262,7 @@ void DLB_MPI_Allreduce_enter (void* sendbuf, void* recvbuf, int count, MPI_Datat
 	fprintf(stderr," >> MPI_Allreduce...............\n");
 	#endif
 	
-	before_mpi(Allreduce, (int)sendbuf, (int)recvbuf);
+	before_mpi(Allreduce, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Allreduce_leave (){
@@ -357,7 +358,7 @@ void DLB_MPI_Bcast_enter (void *buffer, int count, MPI_Datatype datatype, int ro
 	fprintf(stderr," >> MPI_Bcast...............\n");
 	#endif
 	
-	before_mpi(Bcast, (int)buffer, root);
+	before_mpi(Bcast, (intptr_t)buffer, root);
 }
 
 
@@ -374,7 +375,7 @@ void DLB_MPI_Alltoall_enter (void *sendbuf, int sendcount, MPI_Datatype sendtype
 	fprintf(stderr," >> MPI_Alltoall...............\n");
 	#endif
 	
-	before_mpi(Alltoall, (int)sendbuf, (int)recvbuf);
+	before_mpi(Alltoall, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Alltoall_leave (){
@@ -390,7 +391,7 @@ void DLB_MPI_Alltoallv_enter (void *sendbuf, int *sendcnts, int *sdispls, MPI_Da
 	fprintf(stderr," >> MPI_Alltoallv...............\n");
 	#endif
 	
-	before_mpi(Alltoallv, (int)sendbuf, (int)recvbuf);
+	before_mpi(Alltoallv, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Alltoallv_leave (){
@@ -406,7 +407,7 @@ void DLB_MPI_Allgather_enter (void *sendbuf, int sendcount, MPI_Datatype sendtyp
 	fprintf(stderr," >> MPI_Allgather...............\n");
 	#endif
 	
-	before_mpi(Allgather, (int)sendbuf, (int)recvbuf);
+	before_mpi(Allgather, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Allgather_leave (){
@@ -422,7 +423,7 @@ void DLB_MPI_Allgatherv_enter (void *sendbuf, int sendcount, MPI_Datatype sendty
 	fprintf(stderr," >> MPI_Allgatherv...............\n");
 	#endif
 	
-	before_mpi(Allgatherv, (int)sendbuf, (int)recvbuf);
+	before_mpi(Allgatherv, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Allgatherv_leave (){
@@ -438,7 +439,7 @@ void DLB_MPI_Gather_enter (void *sendbuf, int sendcnt, MPI_Datatype sendtype, vo
 	fprintf(stderr," >> MPI_Gather...............\n");
 	#endif
 	
-	before_mpi(Gather, (int)sendbuf, root);
+	before_mpi(Gather, (intptr_t)sendbuf, root);
 }
 
 void DLB_MPI_Gather_leave (){
@@ -454,7 +455,7 @@ void DLB_MPI_Gatherv_enter (void *sendbuf, int sendcnt, MPI_Datatype sendtype, v
 	fprintf(stderr," >> MPI_Gatherv...............\n");
 	#endif
 	
-	before_mpi(Gatherv, (int)sendbuf, root);
+	before_mpi(Gatherv, (intptr_t)sendbuf, root);
 }
 
 void DLB_MPI_Gatherv_leave (){
@@ -470,7 +471,7 @@ void DLB_MPI_Scatter_enter (void *sendbuf, int sendcnt, MPI_Datatype sendtype, v
 	fprintf(stderr," >> MPI_Scatter...............\n");
 	#endif
 	
-	before_mpi(Scatter, (int)sendbuf, root);
+	before_mpi(Scatter, (intptr_t)sendbuf, root);
 }
 
 void DLB_MPI_Scatter_leave (){
@@ -486,7 +487,7 @@ void DLB_MPI_Scatterv_enter (void *sendbuf, int *sendcnts, int *displs, MPI_Data
 	fprintf(stderr," >> MPI_Scatterv...............\n");
 	#endif
 	
-	before_mpi(Scatterv, (int)sendbuf, root);
+	before_mpi(Scatterv, (intptr_t)sendbuf, root);
 }
 
 void DLB_MPI_Scatterv_leave (){
@@ -502,7 +503,7 @@ void DLB_MPI_Scan_enter (void *sendbuf, void *recvbuf, int count, MPI_Datatype d
 	fprintf(stderr," >> MPI_Scan...............\n");
 	#endif
 	
-	before_mpi(Scan, (int)sendbuf, (int)recvbuf);
+	before_mpi(Scan, (intptr_t)sendbuf, (intptr_t)recvbuf);
 }
 
 void DLB_MPI_Scan_leave (){
